+++ src/lang/tr.ts
@@ -2,8 +2,7 @@ import {Lang} from "./Lang";
 
 export const tr: Lang = {
     code: 'tr',
-    display: 'Turkish',
-    flag: '🇹🇷',
+    display: 'Türkçe',
     numbers: [
         'sıfır',
         'bir',